requirements.txt
azure-identity
openai
python-dotenv